+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Translations for relative time
 - Account picker list component for use across the app
 - Account picker on login page, so if you add an account and change your mind or log out you can open your other accounts easily
+- Existing account check & handling to login
 
 ### Fixed
 - Awkward naming of SnackbarController, renamed it LocalSnackbarController